Fix TO CHAR! of BINARY!, expose UTF-8 trailing bytes count
There was a bug when a multi-byte BINARY! was converted to a CHAR!,
because it decremented the number of expected bytes before calling
a service routine instead of after.  The service routine assumed the
total number of bytes was then insufficient for the conversion.

Also exposes a table of trailing byte counts for the leading byte of
a UTF-8 sequence.  Since we have the data inside Rebol, it is useful
to any usermode code that is involved in manually determining how many
bytes to read out of a buffer (prior to passing it to Rebol for the
actual conversion).