Add "include guard" to rebol.h
Generally speaking, the Rebol core does not use include guards in its
implementation.  This offers some amount of guidance on layering the
system in a clearer way--where one piece builds on another without
the lower pieces reaching back and building on the higher piece.

However, libRebol is a single-file include targeting end users.  It
is standard for C library headers to permit multiple inclusions.  So
this adds an "include guard" specifically to rebol.h.