Transition Windows to per-key console implementation
The monolithic ReadConsole() call with ENABLE_LINE_INPUT meant that
the console on Windows was completely blocking--no custom processing
could be done on individual keypresses.  This ruled out things like
tab completion, or even customizing the history mechanism.

Switching over to a per-key-processing implementation represents a
quantum jump in complexity.  But it is necessary to do interesting
features, and the hope is to build a common protocol so that console
code like history and completion can be shared between the web builds
and the Linux/Mac/PC builds.  This leverages the work of factoring the
POSIX console (which chose to implement its own per-key processing)
in a way to share some code.

It still is very early in its development, and needs more work.  But
this shows basic operation.

Writing the Windows version required a new strategy for doing the
buffered TEXT!-input, because it was less obvious when events would
be non-blocking vs. have more printable characters.  This patches
that approach back to POSIX.

This also tries to curb the generic stdio port to not do "device layer"
READ calls when a terminal is in place.  This deference of stdio to
the terminal might be better in another place.