Commit to C99 or above for smart terminal features
With Rebol taking more responsibility for the implementation of the
cursoring and per-key terminal functionality of the Windows console,
an increasing amount of libRebol code is used to implement the smart
console.  libRebol is much more readable and manageable when it can
use the REBOL_IMPLICIT_END functionality.  But that depends on variadic
macros, which requires C99 or higher.

This commits to the idea that if you want rich console features (like
history and tab completion) that the code implementing it is only
available in C99 builds or higher.  This embraces a general philosophy
that while C89 builds are still going to be kept working on Travis
for the core, advanced features (ODBC, ZeroMQ, smart console) will
use the more streamlined form of libRebol.