Fix null redirection on Windows
When using CALL with piping set to FALSE, the idea was that it would
be equivalent to a /dev/null redirect.  This had been tested and
seemed to be working on POSIX, but never appears to have actually been
tried on Windows.

In addition to researching the undocumented method required to do this,
this commit also factors out the /ERROR and /OUTPUT code for
interpreting the argument value into a common routine--due to how
similar they were.