Abandon smart console code if either input/output redirected
The previous concept of abandoning the smart console was that you
would abandon it if the input were redirected, but not the output.
This made some sense when Rebol wasn't taking responsibility for
every keystroke and managing the cursor position.  But now it would
mean managing a dual track...realizing that the notion of where you
write for the console could be different than where you write for
the standard output.

This is a complex topic that needs more study.  The near term solve is
to give up on smart console if any redirection is used.