Fix dropped chars on Ctrl-V pastes in Windows console
The idea of how buffered input would be done changed which branches
of an if() were responsible for advancing the input record pointer.
This resulted in dropped input records due to inconsistent handling.
The effect was noticeable on pastes when done by means of ctrl V
as opposed to menu commands, when the Ctrl key being held down would
result in records being dropped.

This fixes that, as well as moves the responsibility for deciding that
ESCAPE should flush input events out of the key event processing and
into the higher-level readline processing.