Switch READ-STDIN to READ-LINE
A source of concern in the Rebol world for those trying to use it with
CGI scripts is that the lack of a byte-level agnostic way of doing
input and output with redirected I/O made some things impossible.
This can manifest in many ways...like only allowing printing of UTF-8
legal strings, or filtering carriage returns out of the standard
input (which might represent byte values that are not carraige return
in non-textual mediums).

This changes the point at which carriage return filtration was done
away from something called READ-STDIN and now called READ-LINE.  The
idea being that STDIN will evolve into a byte-level agnostic PORT!
that will have some kind of layers on top of it (terminal/console, etc.)
providing higher level services.