Change CALL to INSERT, Windows TEXT! has no CR
Rebol2's CALL apparently inserted into the string or binary you would
pass it for /OUTPUT or /ERROR.  This provides more efficiency than
merely allowing appending, because it cuts down on the number of
intermediate series if you didn't want to add to the end.

This also changes it so that if your /OUTPUT or /ERROR in Windows is
a TEXT!, then you will not get CR in your strings.  If you use a
BINARY!, then you will get whatever bytes are sent across the pipe
literally...which would include the CR bytes if you are running
something like DIR for the shell.  (Note Windows does not use UCS2
for redirection/pipes, only consoles.)

Further improving efficiency, this switches to the use of rebAlloc()'d
buffers for CALL.  These buffers can then be reclaimed directly to use
as BINARY! values, instead of needing to make yet-another-copy.