Tighten series termination on expansion invariants
Early in Ren-C development, there were several questions surrounding
the inconsistent termination of series.  R3-Alpha desired to avoid
unnecessary termination on in-progress series that would be overwritten
at the cost of leading to many situations where it wasn't clear
whether an operation terminated or not, leading to termination "just
to be safe".  In particular, BINARY! series were sometimes terminated
and sometimes not.

Ren-C mandates that BINARY! series have a zero byte terminator, even
though it is past the content.  This makes it possible to decide to
alias a binary as an ANY-STRING! (assuming its bytes are legal UTF-8).

However, since it is dependent on the series subclass to manage this
invariant, the generic routine "Expand_Series" was a bottleneck.  It
was doing a memmove() operation over some of the "used" range of the
available data without necessarily knowing if there was a terminator
implied by that range.

This attempts to tighten those invariants a bit in the hopes that
still more clarity will emerge about how to best express these
core functions.