Add speculative Emoji support to Windows console
Windows's "Command Prompt" is a very old program that does not handle
high unicode codepoints (like Emoji) in rendering, or via its built-in
"Paste" menu code.  (Menu actions are translated into key events by
way of the default handler for the menu item, which we use in order
to avoid implementing all the cut/copy/paste functionality ourselves...
so the omission of Emoji from those pastes is collateral damage.)

However, the insertion of Emoji does seem to work from emoji-enabled
keyboards, such as the virtual keyboard on tablets (not the on-screen
keyboard application in Windows 10...rather the service you get
specifically when you enter tablet mode and push the "keyboard" icon
in the task bar).  That was enough to do a proof-of-concept for emoji
support in Rebol's console.

"Windows Terminal" is an app-store app in upcoming releases which is
purported to handle Emoji correctly, so this should be a step in
preparing for that.