Fix character POKE-ing and SET-PATH!-ing
The original UTF-8 Everywhere implementation was massive in scope, and
not everything could be tested completely before committing it.  This
was considered an acceptable risk to begin reshaping the system around
the BINARY!/ANY-STRING!/ANY-WORD! aliasing implications.

So passing the existing tests was used as an exercise of what was
actually being used.  Rather than preemptively fix all features, the
tests would be allowed to arise organically based on a "squeaky wheel"
theory--that anything actually getting used would get a test.

This addresses overlooked "character poking" features where an
enumeration was not being used.  The SET_CHAR_AT() was not doing its
math correctly, and also was not updating bookmarks.

In addition to fixing the code, this includes a stress test.  It also
decoupled the UTF-8 Encoding routine from determining the size of a
codepoint...which made several "guessed" overestimated string routines
be based on the actual encoding size.  This allowed SET_CHAR_AT() to
avoid redundantly calculating the size of the expected codepoint.