Remove workaround for emscripten compilation bug
There was an issue with compiling in emscripten that was present in
an older version, which forced the rewriting of a branch to repeat a
variable and return statement instead of factoring it out.

Since the time this was a problem, a new version of Clang is now used,
and the compilation methodology has undergone a significant change
(weaving emscripten workarounds into the code generation process via
a Clang API instead of running it as a post-processing pass).  This
particular problem seems to no longer be an issue.