Switch emscripten builds to Linux instead of OS X
When Emscripten builds were initially tried on Travis, the Linux
version did not work.  However it is preferable to OS X for all manner
of things--free virtualization of instances, open source, more hacker
friendly--so this tries to switch it back since a couple years pased.

(The motivation was the pain in getting firefox to be installed and
run from the command-line, but there are other reasons.)