Pass processed window.location.search as system/options/args
The %load-r3.js is JavaScript code which does filtering and processing
of the part of the URL that comes after the `?`...including decoding
any URI components.  This shouldn't error if it sees things it does
not recognize, but rather pass them on to the script (e.g. the REPL)
to see if it understands them.

It makes sense that this would correspond to system/options/args, as
it kind of the equivalent to a command line.  (system/script/args would
be a further processing of arguments sent to a script).

It's not clear exactly how to pass on the arguments, so this uses a
BLOCK! with SET-WORD! and value pairs as a placeholder for thinking
about a more general concept.  But it simplifies the code in replpad-js