ENBIN + DEBIN draft/sketch implementation
Rebol2 had an asymmetrical sense of conversion for BINARY! and INTEGER!

    rebol2>> to binary! 32
    == #{3332}  ; #{33} is ASCII "3", #{32} is ASCII "2"

    rebol2>> to integer! #{3332}
    == 13106  ; 0x3332 is the internal big endian form of 13106

R3-Alpha "corrected" this unusual behavior by changing TO BINARY!.  The
conventional wisdom seemed to be that `TO BINARY! TO STRING! 32` was
not a common desire, but easy enough to express if you wanted it.  The
harder conversion for users (which was "easy" for Rebol to do) involved
the internal byte representations of native C integers:

    r3-alpha>> to binary! 32
    == #{0000000000000020}  ; 0x20 is 32 in hexadecimal

    r3-alpha>> to integer! #{0000000000000020}
    == 32

While this might seem more sensible on the surface, it was awkward for
users to get the INTEGER! <> BINARY! conversions they wanted...where
details of signedness, byte-size, and byte ordering vary considerably.
Users might want #{FF00} to mean little-endian 255, big-endian
unsigned 65280, or big-endian signed -256.  Getting these results from
a fixed 8-byte signed big-endian conversion was hard and error prone.

An added concern is that Ren-C's philosphy is that a futureproof
language at Rebol's level of concern should always work with integers
that can scale to arbitrary-precision.  The goal is to finesse the
nature of immutable and mutable INTEGER! in the design so that when
immutable integers fit in a value cell, they can be operated on
efficiently...then only mutable integers would need BigNum "nodes"
pointed to by the cell.  If care is taken to make immutability the
common case, this could provide the best of both worlds.

Yet when Ren-C tried out an unusual way to make TO BINARY! of an
INTEGER! return a variable number of bytes, that didn't wind up
helping much.  This commit introduces two new tools--tentatively
called ENBIN and DEBIN--which take a BLOCK! describing the encoding
or decoding of a binary to integer that is desired:

    >> enbin [be + 4] 32
    #{00000020}  ; big-endian, 4 byte, unsigned

    >> enbin [le + 2] 32
    #{2000}  ; little-endian, 4 byte, unsigned

    >> enbin [le +/- 3] -2
    == #{FEFFFF}  ; little-endian, 3 byte, signed

    >> enbin [le + 3] 16777214
    == #{FEFFFF}  ; little endian 3 byte, this time unsigned

    >> enbin [le +/- 3] 16777214
    ** Error: 16777214 aliases a negative value with signed
           encoding of only 3 bytes

DEBIN is ENBIN's complement which uses the same dialect but goes the
other way:

    >> debin [le + 3] #{FEFFFF}
    == 16777214  ; reverse of above example

The dialect choice was made with the endianness first to possibly
suggest compatibility with WORD!-based codecs and ENCODE/DECODE if they
took a block (encode/decode are nicer words then enbin/debin, but this
is just a test for now so not conflating them).  The sign is in the
middle because unlike ENBIN, DEBIN can guess the size accurately from
the length of the input.

This is a prelude to possibly walking back the TO BINARY! of an integer
semantics to Rebol2 and making TO INTEGER! of a BINARY! reverse *that*.

https://forum.rebol.info/t/1270