Switch some user-config driven cflags to be per-file
The Rebol-based build system from Atronix has two main parts.  One
is something called "Rebmake", which models parts of the build system
as objects (e.g. obj files, EXEs, linkers...)  This could theoretically
be used for any build tooling you wanted to do--though making it useful
for anything but building Rebol is not any kind of near-term goal.

The other is a build script specific to building Rebol itself.  That's
%make.r in the root directory of the project.  It translates its file
lists and extensions into Rebmake objects.  So one of its jobs is to
take Rebol blocks full of file lists and tags describing any special
handling that file needs (e.g. if it's a third-party file and doesn't
heed warnings that would be corrected in files Ren-C controls)

However %make.r was trying to generate a list of flags for the project
based on settings in the user config (such as whether the project
should be built as C++ or C, translating into a complex set of flags).
This flag list was hard to amend after-the-fact...such as to point out
that a certain third-party file wasn't able to be built as C++.

This hacks in the generation process for more of those flags coming
from the user config into the OBJ generation function.  It adds a
slight cost to the build time to reprocess it for each file...but that
is irrelevant given the increase in flexibility.