Switch RC4 encryption to use mbedTLS
The RC4 encryption method is generally deprecated, but code to support
it existed in Saphirion's build.  Even though it is no longer used in
TLS cipher suites we support, the code had not been deleted from the
Crypt module.  It serves as an example and test of how to write code
of that form...and someone might find it useful.

This replaces the old axTLS rc4 code with the more maintained mbedTLS
version.  Ultimately it should likely not be included by default, but
as with most crypto algorithms be its own selectable extension.