Workaround gcc=g++ setting in configs for #no-cpp
The #no-cpp setting appeared to work in Visual Studio because the
settings for C++ compilation were entirely on the command line for
the same compiler executable.  But the "generic-c++.r" configuration
used `g++` as the compiler instead of `gcc`...which meant that there
would have to be a special switch used to turn C++-ness *off* (vs
relying on switches to turn it on).

This applies a workaround by which if a compilation was deemed to be
C++ for a file then it will add the TAG! `<gnu:-x c>`, which is a way
of saying that a switch applices for GCC only.  There are surely more
elegant ways to do this but right now the task at hand seems to be
served by it (pending greater design review of Rebmake).