Move and stylize random number generator for mbedTLS
Because there are many different ways to make random numbers based
on "available sources of entropy", the mbedTLS library does not
build in assumptions about random number generators...and any routine
that requires randomness takes a generator function.

The Saphirion cryptography used /dev/urandom on POSIX and on Windows
used CryptGenRandom().  Rather than rock the boat at the moment on
trying to do better at the cost of potentially more code, we stick
with that.

This moves the code for random generation out of the non-mbedTLS
rsa.c (which will be going away in a subsequent commit in favor of
using the mbedTLS one).  It also changes the parameterization to
match the expectations of mbedTLS for a random number generator...
which affords the ability at callsites to effectively implement
a closure by tunneling in a pointer parameter.