Add extra underscore to generated parameter name
If you have an argument called `SIZE` to a native, and you used the
parameter generating macros, you would get an argument named `p_size`.
This is actually a fairly reasonable name for a variable (and it did
come up in the crypto, because there were P (modulus) and G (base)
and then a `p_size`.

To make the variable something even less likely to type, this glues
another underscore onto it so you get `p_size_` in such a case.