Message Digest abstraction layer plus SHA512, RIPEMD160
A previous commit converted the SHA256 primitive to use mbedTLS instead
of a one-off sha256.c file from axTLS.

This commit builds on that successful migration by using the message
digest abstraction layer offered by mbedTLS.  This builds a table of
the hashes you choose to use from the config that can be listed and
queried at runtime, and gives a common interface to using any of those
hashes--with or without a HMAC key.

This allowed dropping the MD5 and SHA1 implementations which were
being used, which had been factored into an attempted abstraction
layer that was similar in spirit--which powered the CHECKSUM command.
Because doing such factoring involved effort, some message digests
(such as 16-bit TCP checksums) were simply added as refinements--
even though that didn't make much sense in combination with other
parameters.

Beyond offering more optimized/clear/peer-reviewed versions of the
digests Rebol historically had, this also opened the doors to including
ones it didn't--like SHA512 and RIPEMD160.

Additionally, the design of the checksum interface in mbedTLS permits
streaming checksums, so that this implementation could be progressed
into a PORT!-like model which could do things like checksum an entire
.iso file for a CD image without needing to actually load the whole
file in at once, but gradually feed chunks of it through the hashing.

This meant that the previous commit's work on including SHA256 gets
subsumed by the more general method.  It also means that a usermode
function for generating a SHA256 with HMAC exists as well.  Rather
than throw away the usermode version entirely, this moves it into
a test file to help exercise more usermode code against a known goal.