Limit platforms for type-punned structs for header
The bitfield-based type punning to see individual flags in a header
in some debug builds is non-standard, and has caused problems on
platforms used in the past.  However, it is quite helpful when one
needs it.

It turns out this is causing trouble in cross-compiled builds on
some Linux x64 platforms when building for Linux x86.  The feature is
now limited to being by default as in those debug builds solely for
little-endian Linux x64, which has yet to demonstrate problems.