Use Valgrind on non-address-sanitizer builds
Some builds are done with DEBUG=none to make sure they work.  However,
if they don't work the crash messages aren't very helpful.

Usually this can be debugged by virtue of doing the build locally.
However, the cross-compiled x86 build on Linux has on at least one
occasion failed on Travis but not locally with the same settings.

To better understand what is going on in such cases, this installs
Valgrind...which can give more meaningful output on a non-debug crash,
assuming the executable was built with symbols.

(Note: Valgrind cannot be used on ASAN builds.)