Try switching Linux x86 build to bionic
There appears to be a problem in the cross-compiling toolchain
with inlining in the TCC extension.  Looking into the issue only
seems to occur on the Travis container, and inlining the function
manually seemed to be fine.  There does not appear to be any
violations of C rules at play, and updating the compiler resolves
the issue--suggesting strongly that it is indeed a bug in that
version of the cross-compilation toolchain (which tends to lag
in getting fixes compared to the native toolchains).