BLANK! as space character in DELIMIT/UNSPACED/SPACED
This embraces the concept of "what space is to strings, BLANK! is to
blocks", by taking that notion forward to let a BLANK! be rendered
as a space character:

    >> unspaced ["a" "b" _ "c"]
    == "ab c"

Having a shorthand notation for a space improves visibility while
saving on characters (over typing `space` out as a word):

    >> unspaced ["x:" space first block space "+ 10"]
    vs.
    >> unspaced ["x:" _ first block _ "+ 10"]

It also can be preferable to tacking spaces in with adjoining string
parts (in cases where that is possible).  Because they are the same
character in the language used to separate values themselves.  This
makes it sometimes hard to see where a string beginning or end is:

    >> unspaced ["x: " first block " + 10"]
    ; ...at a glance, `" first block "` might look like a string
    ; vs.
    >> unspaced ["x:" _ first block _ "+ 10"]

So it offers advantages there too.

This idea had been tried before and rejected in favor of making BLANK!
a synonym for NULL.  The reason was that early designs of NULL unified
it with VOID!, so it was the way of saying a variable was undefined.
This meant that despite its status as the true "non-value" which
couldn't be confused with something you might want to put into a block,
it was "ornery" to deal with.  This led many cases for "opting out" to
favor the idea of opting out with BLANK! and considering NULL an error.

As time went on, BLANK! increasingly seemed to be the "reified NULL"
where NULL was representative of a "soft error" condition that could
be defused by transforming nulls to blanks via TRY.  This furthered
the idea that dialects should likely make the two mean the same thing
(if both nulls and blanks were to be permitted), otherwise error on
NULL and use blanks as their proxy.

Slowly, however, the ornery-ness of NULL was chipped away.  It was
decided to be too convenient to use IF in PRINT statements without
having to "TRY" to defuse the case when the branch did not run and
it produced NULL.  Eventually, this went all the way and NULL no
longer would cause errors when either assigned via SET-WORD! or fetched
via normal WORD! access.

With NULL being able to "mean what it means" as the true absence of
a value, it opens the floor for letting BLANK! mean something else.