Fix bookmark getting out of sync on INSERT operations
The string code was updating the byte offset of a bookmark when the
insertion index occurred earlier in the string than the index position
for the bookmark.  However, it was not updating the index position
of the bookmark itself to account for the number of codepoints that
were being inserted to push it to its new later position.

This adds the number of codepoints (this had already been taken into
account in the CHANGE operation branch).