Shut down core and restart it before shutting down in debug
The theory is that you should be able to start up Rebol, use it,
and shut it down such that it releases its resources...you could
then start it again later.  Being able to do this clean round-trip
of initialization demonstrates a good amount of control over state.

This gives a first attempt at being able to shutdown and start and
shutdown again in the debug build.  This required a small change to
the (largely legacy) device layer to free up any pending events
for the device.