Enable WORDS OF for ACTION!s
When action parameters in the dialect were no longer all specified
purely by words, the PARAMETERS OF operation was introduced to capture
that idea.

However, it is desirable sometimes to get purely a list of words of
an ACTION! as they would appear in the frame.  Previously this had to
be done with MAKE FRAME! and then getting words of that frame.  This
adds the feature for getting the pure words back (without knowing
if they are refinements or not, just that they are named elements of
a frame).