Prototype of Multiple Return Values (slow, conceptual)
This is a deliberately high-level implementation of a multiple return
value concept, using the libRebol API inside of the evaluator
implementation itself.  It is done in such a way in order to facilitate
changes to the concept more easily while the idea is tested for
viability.

The idea is to declare functions with multiple outputs in much the
same way as historical Rebol, where WORD! or PATH! gets passed in to a
routine which is the variable that is then set by the code.  The
code may decide to do additional processing or change its behavior
based on the presence or absence of the request for those variables.

The difference is that it offers the ability to specifically label such
refinements as <output>, and if this is done then it will participate
in the evaluator with the SET-BLOCK! construct, which will do an ordered
injection of parameters from the left-hand side.  It takes care of
pre-composing any PATH!s with GROUP!s in them, as well as voiding the
variables.  This means that while the called routine can vary its
calculations based on whether or not an output is requested, it can't
depend on the previous contents of the variable.

Currently the idea is that the first element of the block represents
a variable that would have been the conventional result of the
operation.  This is because `[a b c]: whatever` seems more elegant
than writing `a: [b c]: whatever`.  Hence the result of the evaluation
is the same thing that is put in the first slot.

This results in functions that can be used in the traditional way or
that can take advantage of the shorthand.