Make %sys-xxx.r from mezzanine scan without LOAD in bootstrap
The STRIPLOAD function had already been made to take out the comments
and whitespace from the mezzanine.  (That had been a previous reason
for forcing the bootstrap executable to LOAD them.)

This made it possible to use new scan constructs immediately without
waiting for the bootstrap Rebol to change.  But there was a lingering
dependency for the %sys-xxx.r files, which was using MAKE OBJECT! to
find all the top level SET-WORD!s in order to create some `#define`
values in a header file for them.

While a heuristic process to analyze the %sys-xxx.r files for what
look like top-level SET-WORD!s is not going to be perfect, the files
follow a fairly regimented pattern that can be limited to modifications
by people who understand the rules.  Additionally, the debug build
can catch the case where the helper function numbers get out of sync
by checking a string-based ID matched against the context key.

This implements the approach in order to liberate the usage of new
scanner conventions in the %sys-xxx.r files--e.g. the LOAD routines.