Tolerate warnings when building for XP with VS2019
Visual Studio 2019 has deprecated building with the platform library
that can run on Windows XP.  You can still install the VS2017
component to do it (for now).  But if you do, then the newer compiler
will give warnings about things that it doesn't like about the old
platform SDK header files.

For the sake of knowing what the issues are, this disables those
warnings in %make.r, so that a rigorous build can still succeed when
targeting XP.