Make multi-returns void *after* function is running
The initial prototype of multiple returns would build a complicated
proxy function out of a specialized enclosure.  The specialization
would handle poking the variables to set into their respective output
arguments.  And the enclosure would tie it up at the end so that the
first element of the block would receive a copy of the ordinary result.

For safety, variables in the block were set to void before running it.
This caught functions that didn't assign their outputs, as well as
stopping the function from treating variables on the left as input
arguments, reacting to the existing state of whatever was in it.

However the voiding was done in the specialization phase, before the
function was run through a re-evaluation to gather arguments.  That
meant you couldn't use a variable on both the left and right hand
sides, because the right hand side would see it as already voided.

This moves the voiding into the enclosure so it happens after the
arguments have been evaluated, but before the function runs.