TRANSCODE redesign with multi-return, error improvements
Much of LOAD was written in userspace and built on top of TRANSCODE.
In R3-Alpha, this meant the scanner was limited in what it could
report for errors...it was receiving strings and knew what line on
the string it was on, but did not know the filename to report any
errors on.  Also, it only knew the line offset within the string...
so when headers were scanned separately the line numbers were off
by the number of lines in the header.

Ren-C had tunneled in a line number variable to TRANSCODE to resolve
the line number issue.  However, the filename was not threaded through
the loading machinery...which was difficult to navigate and read.

This commit attempts to get to the bottom of the issue by making LOAD
account for the filename.  As part of the process, this involves
improving LOAD and LOAD-HEADER to use a revamped variation of
TRANSCODE which takes advantage of multiple return values.

    value: transcode data  ; transcodes the entire thing
    [value pos]: trancode data  ; transcode one value, give next pos
    [value pos err]: transcode data  ; same but recover from errors

Error recovery (the TRANSCODE/RELAX option) was never very well
mapped out in R3-Alpha... nor was the /ONLY option.  The intended
feature is presumably some kind of "recoverable" scan--which cannot
be done with multiple calls to TRANSCODE when the scanning state has
been abandoned.  A deeper design for that would be necessary (and
would be desirable to have in PARSE as well; to suspend and resume
an in progress parse stack).

However, this preserves a variant of the /RELAX feature by allowing
recovery as long as the error is at the outermost depth level of
the transcode.  It repairs an issue where deeper errors in the scan
were being casually suppressed in an undetectable way, leading to
confusing or absent error reports when /RELAX was used.