Change LOAD to use multi-return semantics
R3 historically had a LOAD which would change what it returned on
the basis of a /HEADER refinement.  If you used that refinement, then
instead of just getting the body of loaded code you would get a block
whose second item was the header, and whose first was a block of
the rest of the data.  If no header was present, the second item would
be a NONE!.

This shifts Ren-C to use the multi-return semantics, where /HEADER
takes an argument which is a WORD! or PATH! to set the header to.
Then if a SET-BLOCK! is used on the left, then if it has two elements
then the second element will be set to the header's value.  The
overall return--as well as the first element--will be set to the
data for the LOAD>

As a result, when a header is not present the returned value for the
header will be NULL.  This simplifies working with code that wants to
process the header separately from the data.

Note: Because SET-BLOCK! interactions with FUNCTION or LET in FUNC
have not been completely hammered out, variables that need to be
"gathered" are often moved to become local prior to usage in the
SET-BLOCK!.  This is a temporary state of affairs, as resolution on
how this will be approached gets worked out.