Fix C99 or higher Windows builds using smart console
The intent was that C99 or higher builds or C++11 or higher builds
would get smart console features, disabling them for older compilers
in order to allow the source to be more elegant via use of variadic
macros.  However, the #ifdefs were written along the lines of:

     if (using C-compiler and not at least C89) {
          // don't use smart console (no variadic macros)
     }
     else if (not a C++ compiler that's at least C++11) {
         // don't use smart console (no variadic macros)
     }
     else {
          // use smart console
     }

This meant no C compiler at any version would make it to the smart
console branch when written that way.  The second `if` needed to say
`if (using C++ compiler and not at least C++11)`.

Additionally, it turns out that some C compilers do not define
__STDC_VERSION__ at all.  The suggestion of defining it post-dates C89
itself, so even MinGW building as —std=gnu89 doesn’t retroactively
define it.  Cover this case so it doesn’t try to build with a smart
console enabled.