Work around ERROR_NOT_ENOUGH_MEMORY on console read
Some versions of Windows, notably Windows 7, have trouble with reading
“large” size requests using ReadFile() on the standard input handle
when that input handle is connected to a console.  The Go language
noticed this problem also:

https://github.com/golang/go/issues/13697

The situation arose because Ren-C is now using ReadFile() instead of
ReadConsole() to get input not just when stdin is redirected, but also
when the build is older than C99 or C++11.  This is because the granular
console code is more complex and looks better when variadic macros can
be used; and it is assumed any practical system that has smart terminal
features probably also has a C99 compiler one could use on it.

Hence the issue was noticed on Windows 7.  To remedy it, the size of
the buffer is backed off incrementally and the read re-tried.

To make sure this branch gets tested, the x86 Windows build is chosen
to be built as C89 with no smart console enabled.